Move SDKMethodRunner to top level
